+++ resources/js/Pages/English/Wishes.vue
@@ -1,0 +1,8 @@
+<script setup>
+import GrammarLesson from '../../Components/GrammarLesson.vue';
+import content from '../../../content/english/wishes.json';
+</script>
+
+<template>
+    <GrammarLesson :content="content" />
+</template>